Move example to another module.
