+++ drfs/structure.py
@@ -1,9 +1,9 @@
-from textwrap import indent
 import inspect
 from pathlib import Path
+from textwrap import indent
+from typing import Union
 
 from .path import DRPath
-from typing import Union
 
 
 class _MetaTree(type):
@@ -72,59 +72,9 @@ class _BaseTree:
 
 
 class Tree(_BaseTree, metaclass=_MetaTree):
-    """Subclass from this to create your file structure."""
+    """Subclass from this to create your file structure. See _example.py."""
     pass
 
 
+# Use this as a type hint for paths for better autocomplete.
 P = Union[Path, DRPath]
-
-
-# ------- EXAMPLE --------
-
-
-class _ExampleStructure(Tree):
-    # This is the simplest way of defining a path.
-    raw = 'raw'
-    
-    # Use P type hint for better autocomplete. Use DRPath type to reuse path.
-    data: P = DRPath('data')
-    
-    # You can reuse paths defined earlier as DRPaths.
-    processed = data / 'processed'
-    
-    # Of course, you can also have templates and wildcards.
-    dataset = processed / '{date}' / '*'
-    
-    # Nest another tree to group paths. Name of tree will be used as parent
-    # folder.
-    class migrated(Tree):
-        things: P = 'thngs'
-        stuff: P = 'stf'
-    
-    # You can override name of parent folder by defining __root__ attribute.
-    class report(Tree):
-        __root__ = 'rprt'
-        evaluation = 'eval'
-
-
-# Instantiate the defined class giving it a name of root directory.
-_example_structure = _ExampleStructure('ROOT')
-# You can update root directory at any time. All paths will reflect this.
-_example_structure.root = 'NEW_ROOT'
-
-# This is what the structure looks like now:
-"""
-_root: NEW_ROOT
-data: NEW_ROOT/data
-dataset: NEW_ROOT/data/processed/{date}/*
-migrated:
-    _root: NEW_ROOT/migrated
-    stuff: NEW_ROOT/migrated/stf
-    things: NEW_ROOT/migrated/thngs
-
-processed: NEW_ROOT/data/processed
-raw: NEW_ROOT/raw
-report:
-    _root: NEW_ROOT/rprt
-    evaluation: NEW_ROOT/rprt/eval
-"""
